Incorporate comments
